// music_app/types/music-video.ts
export interface MusicVideo {
  idTrack: string;
  idAlbum: string;
  strArtist: string;
  strTrack: string;
  intDuration: string;
  strTrackThumb: string | null;
  strMusicVid: string;
  strDescriptionEN: string;
  strMusicBrainzArtistID: string;
  strMusicBrainzAlbumID: string;
  strMusicBrainzID: string;
}